% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visnetwork.R
\name{merge_edges}
\alias{merge_edges}
\title{Merge Duplicate Concept Map Edges}
\usage{
merge_edges(data)
}
\arguments{
\item{data}{(dataframe) A table created from, to, & label columns}
}
\value{

}
\description{
Merge Duplicate Concept Map Edges
}
