% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visnetwork.R
\name{plot_cleaned_visnetwork}
\alias{plot_cleaned_visnetwork}
\title{Plot visNetwork with Merged Edges}
\usage{
plot_cleaned_visnetwork(data)
}
\arguments{
\item{data}{(dataframe) A table created from, to, & label columns}
}
\value{

}
\description{
Plot visNetwork with Merged Edges
}
